Support horizontal scrolling content
We now check for a certain y-axis touch distance and intercept if it
passes the threshold. This allows us to use horizontal scrolling content
such as a RecyclerView with a horizontal LinearLayoutManager inside
VerticalViewPager content.

Threshold may very depending on your layout. For example, if the
VerticalViewPager takes up a large amount of a height, the default
threshold (16dp) works perfectly. But if your VerticalViewPager takes up
less than half the screen height, then a smaller threshold would provide
a better user experience. For that reason, the threshold is customizable
via XML with the vvp_threshold attribute.

This has been tested extensively on a large-scale app that I'm currently
working on.